requirements.txt
aiobotocore
aiofiles
aiokafka
aioredis
ciso8601
cython
fastapi
loguru
motor
multimethod
numpy
orjson
pkgconfig
pymongo
protobuf
pycapnp
pyzmq
sqlalchemy
uuid
uvicorn
uvloop
